requirements.txt
blinker==1.4
boto3==1.24.66
botocore==1.27.66
cachetools==4.2.4
certifi==2022.6.15
cloudpickle==2.1.0
Deprecated==1.2.13
docstring-parser==0.14.1
fire==0.4.0
google-api-core==2.8.2
google-api-python-client==1.12.11
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-storage==1.44.0
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
httplib2==0.20.4
imageio==2.21.2
jmespath==1.0.1
jsonschema==3.2.0
kfp==1.8.13
kfp-pipeline-spec==0.1.16
kfp-server-api==1.8.4
kubernetes==18.20.0
networkx==2.8.6
opencv-python==4.6.0.66
pandas==1.4.2
Pillow==9.2.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pydantic==1.9.2
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.3.0
PyYAML==5.4.1
requests-toolbelt==0.9.1
s3transfer==0.6.0
scikit-image==0.19.3
scipy==1.9.0
strip-hints==0.1.10
tabulate==0.8.10
tifffile==2022.8.12
tqdm==4.64.0
typer==0.6.1
uritemplate==3.0.1
websocket-client==1.3.3